requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psutil==6.0.0
pyyaml==6.0.2
